functionality for PSADT
Added extra check for the case when user wants to pack PSADT folder. The tool first checks if there is an exe file in the App folder, and if yes, it proceeds to the creation of the .intunewin file. If not it gets one level down and prepares the final file based on the Deploy-Application.exe.